admin_bot/requirements.txt
aiogram>=3.10
openai>=1.0
python-dotenv>=1.0
aiohttp>=3.8
asyncpg>=0.28.0
psycopg2-binary>=2.9.7